Start of new bcv resource endpoint
