+++ src/endpoints/git.rs
@@ -9,10 +9,9 @@ use rocket::{get, post, State};
 use rocket::serde::json::Json;
 use rocket::http::{ContentType, Status};
 use rocket::response::status;
-use chrono::prelude::Utc;
 use serde_json::{json, Value};
 use crate::static_vars::NET_IS_ENABLED;
-use crate::structs::{AppSettings, BurritoMetadataIngredient, GitStatusRecord, NewContentForm, NewScriptureBookForm};
+use crate::structs::{AppSettings, BurritoMetadataIngredient, GitStatusRecord, NewScriptureBookForm};
 use crate::structs::BurritoMetadata;
 use crate::utils::files::load_json;
 use crate::utils::json_responses::{make_bad_json_data_response, make_good_json_data_response};
@@ -60,354 +59,6 @@ pub fn list_local_repos(state: &State<AppSettings>) -> status::Custom<(ContentTy
         (
             ContentType::JSON,
             serde_json::to_string(&quoted_repos).unwrap(),
-        ),
-    )
-}
-
-/// *`POST /new`*
-///
-/// Typically mounted as **`/git/new`**
-///
-/// Creates a new, local content repo. It requires the following fields as a JSON body:
-/// - content_name (string)
-/// - content_abbr (string)
-/// - content_type (string)
-/// - content_language_code
-/// - versification (string)
-/// - add_book (boolean)
-/// - book_code (null or string)
-/// - book_title (null or string)
-/// - book_abbr (null or string)
-/// - add_cv (null or boolean)
-#[post("/new",
-    format = "json",
-    data = "<json_form>")]
-pub fn new_repo(
-    state: &State<AppSettings>,
-    json_form: Json<NewContentForm>,
-) -> status::Custom<(ContentType, String)> {
-    // Check template type exists
-    let path_to_template = format!(
-        "{}{}templates{}content_templates{}{}{}metadata.json",
-        &state.app_resources_dir,
-        os_slash_str(),
-        os_slash_str(),
-        os_slash_str(),
-        json_form.content_type.clone(),
-        os_slash_str(),
-    );
-    if !std::path::Path::new(&path_to_template).is_file() {
-        return status::Custom(
-            Status::BadRequest,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response(format!("Metadata template {} not found", json_form.content_type)),
-            ),
-        );
-    }
-    // Build path for new repo and parent
-    let path_to_new_repo_parent = format!(
-        "{}{}_local_{}_local_",
-        state.repo_dir.lock().unwrap().clone(),
-        os_slash_str(),
-        os_slash_str(),
-    );
-    let path_to_new_repo = format!(
-        "{}{}{}",
-        path_to_new_repo_parent.clone(),
-        os_slash_str(),
-        json_form.content_abbr.clone()
-    );
-    // Check path doesn't already exist
-    if std::path::Path::new(&path_to_new_repo).exists() {
-        return status::Custom(
-            Status::BadRequest,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response(format!("Local content called '{}' already exists", json_form.content_abbr)),
-            ),
-        );
-    }
-    // Make parents?
-    match std::fs::create_dir_all(path_to_new_repo_parent) {
-        Ok(_) => (),
-        Err(e) => return status::Custom(
-            Status::InternalServerError,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response(format!("Could not create local content directories: {}", e)),
-            ),
-        )
-    }
-    // Init repo
-    let new_repo = match Repository::init(&path_to_new_repo) {
-        Ok(repo) => repo,
-        Err(e) => return status::Custom(
-            Status::InternalServerError,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response(format!("Could not create repo: {}", e)),
-            ),
-        )
-    };
-    // Set up local user info
-    let mut config = new_repo.config().unwrap();
-    config.set_str("user.name", whoami::username().as_str()).unwrap();
-    config.set_str("user.email", format!("{}@localhost", whoami::username().as_str()).as_str()).unwrap();
-    // Make ingredients dir
-    let path_to_ingredients = format!(
-        "{}{}ingredients",
-        path_to_new_repo,
-        os_slash_str(),
-    );
-    match std::fs::create_dir(&path_to_ingredients) {
-        Ok(_) => (),
-        Err(e) => return status::Custom(
-            Status::InternalServerError,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response(format!("Could not create ingredients directory for repo: {}", e)),
-            ),
-        )
-    }
-    // Read and customize metadata
-    let mut metadata_string = match std::fs::read_to_string(&path_to_template) {
-        Ok(v) => v,
-        Err(e) => return status::Custom(
-            Status::InternalServerError,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response(format!("Could not load metadata template as string: {}", e)),
-            ),
-        )
-    };
-    let now_time = Utc::now();
-    metadata_string = metadata_string
-        .replace("%%ABBR%%", json_form.content_abbr.as_str())
-        .replace("%%CONTENT_NAME%%", json_form.content_name.as_str())
-        .replace("%%CONTENT_NAME%%", json_form.content_name.as_str())
-        .replace("%%CREATED_TIMESTAMP%%", now_time.to_string().as_str());
-    // Get versification file as JSON
-    let path_to_versification = format!(
-        "{}{}templates{}content_templates{}vrs{}{}.json",
-        &state.app_resources_dir,
-        os_slash_str(),
-        os_slash_str(),
-        os_slash_str(),
-        os_slash_str(),
-        json_form.versification.clone().unwrap(),
-    );
-    let versification_schema = match load_json(&path_to_versification) {
-        Ok(j) => j,
-        Err(e) => return status::Custom(
-            Status::InternalServerError,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response(format!("Could not load versification JSON: {}", e)),
-            ),
-        )
-    };
-    // Write it out to new repo
-    let path_to_repo_versification = format!(
-        "{}{}ingredients/vrs.json",
-        path_to_new_repo,
-        os_slash_str(),
-    );
-    let versification_string = serde_json::to_string(&versification_schema).unwrap();
-    match std::fs::write(path_to_repo_versification, &versification_string) {
-        Ok(_) => (),
-        Err(e) => return status::Custom(
-            Status::InternalServerError,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response(format!("Could not write versification to repo: {}", e)),
-            ),
-        )
-    }
-    // Make new book if necessary:
-    if json_form.add_book {
-        let scope_string = format!("\"{}\": []", json_form.book_code.clone().unwrap().as_str());
-        metadata_string = metadata_string.replace("%%SCOPE%%", scope_string.as_str());
-        // - Read and customize USFM template
-        let path_to_usfm_template = format!(
-            "{}{}templates{}content_templates{}{}{}book.usfm",
-            &state.app_resources_dir,
-            os_slash_str(),
-            os_slash_str(),
-            os_slash_str(),
-            json_form.content_type.clone(),
-            os_slash_str(),
-        );
-        let mut usfm_string = match std::fs::read_to_string(&path_to_usfm_template) {
-            Ok(v) => v,
-            Err(e) => return status::Custom(
-                Status::InternalServerError,
-                (
-                    ContentType::JSON,
-                    make_bad_json_data_response(format!("Could not load USFM template as string: {}", e)),
-                ),
-            )
-        };
-        usfm_string = usfm_string
-            .replace("%%BOOKCODE%%", json_form.book_code.clone().unwrap().as_str())
-            .replace("%%BOOKNAME%%", json_form.book_title.clone().unwrap().as_str())
-            .replace("%%CONTENTNAME%%", json_form.content_name.clone().as_str())
-            .replace("%%BOOKABBR%%", json_form.book_abbr.clone().unwrap().as_str());
-        // - If ve
-        if json_form.add_cv.unwrap() {
-            // Generate cv USFM
-            let mut cv_bits = Vec::new();
-            let versification_ob = match &versification_schema {
-                Value::Object(o) => o,
-                _ => return status::Custom(
-                    Status::InternalServerError,
-                    (
-                        ContentType::JSON,
-                        make_bad_json_data_response(
-                            format!(
-                                "Could not find versification JSON object for {}",
-                                json_form.versification.clone().unwrap()
-                            )
-                        ),
-                    ),
-                )
-            };
-            let max_verses_ob = match &versification_ob["maxVerses"] {
-                Value::Object(o) => o,
-                _ => return status::Custom(
-                    Status::InternalServerError,
-                    (
-                        ContentType::JSON,
-                        make_bad_json_data_response(
-                            format!(
-                                "Could not find maxVerses in versification JSON for {}",
-                                json_form.versification.clone().unwrap()
-                            )
-                        ),
-                    ),
-                )
-            };
-            let book_max_verses_arr = match &max_verses_ob[&json_form.book_code.clone().unwrap()] {
-                Value::Array(a) => a,
-                _ => return status::Custom(
-                    Status::InternalServerError,
-                    (
-                        ContentType::JSON,
-                        make_bad_json_data_response(
-                            format!(
-                                "Could not find maxVerses for {} in versification JSON for {}",
-                                json_form.book_code.clone().unwrap(),
-                                json_form.versification.clone().unwrap()
-                            )
-                        ),
-                    ),
-                )
-            };
-            let mut chapter_number = 0;
-            for max_verse in book_max_verses_arr {
-                chapter_number += 1;
-                cv_bits.push(format!("\\c {}", chapter_number));
-                cv_bits.push("\\p".to_string());
-                let max_verse_number = max_verse.as_str().unwrap().parse::<i32>().unwrap();
-                for verse_number in 1..=max_verse_number {
-                    cv_bits.push(format!("\\v {}", verse_number));
-                }
-            }
-            // Insert
-            usfm_string = usfm_string
-                .replace(
-                    "%%STUBCONTENT%%",
-                    cv_bits.join("\n").as_str(),
-                );
-        } else {
-            usfm_string = usfm_string
-                .replace(
-                    "%%STUBCONTENT%%",
-                    "\\c 1\n\\p\n\\v 1\nFirst verse...",
-                );
-        }
-        // - add ingredient to metadata
-        let ingredient_json = json!(
-            {
-                format!("ingredients/{}.usfm", json_form.book_code.clone().unwrap()): {
-                    "checksum": {
-                        "md5": format!("{:?}", md5::compute(&usfm_string))
-                    },
-                    "mimeType": "text/plain",
-                    "size": usfm_string.len(),
-                    "scope": {
-                        format!("{}", json_form.book_code.clone().unwrap()): []
-                    }
-                },
-                "ingredients/vrs.json": {
-                    "checksum": {
-                        "md5": format!("{:?}", md5::compute(&versification_string))
-                    },
-                    "mimeType": "application/json",
-                    "size": versification_string.len()
-                }
-            }
-        );
-        metadata_string = metadata_string
-            .replace(
-                "\"ingredients\": {}",
-                format!("\"ingredients\": {}", ingredient_json.to_string().as_str()).as_str(),
-            );
-        // - Write USFM
-        let path_to_usfm_destination = format!(
-            "{}{}ingredients{}{}.usfm",
-            &path_to_new_repo,
-            os_slash_str(),
-            os_slash_str(),
-            json_form.book_code.clone().unwrap(),
-        );
-        match std::fs::write(path_to_usfm_destination, usfm_string) {
-            Ok(_) => (),
-            Err(e) => return status::Custom(
-                Status::InternalServerError,
-                (
-                    ContentType::JSON,
-                    make_bad_json_data_response(format!("Could not write usfm to repo: {}", e)),
-                ),
-            )
-        }
-    } else { // No ingredients
-        metadata_string = metadata_string.replace("%%SCOPE%%", "");
-    }
-    // Write metadata
-    let path_to_repo_metadata = format!(
-        "{}{}metadata.json",
-        &path_to_new_repo,
-        os_slash_str()
-    );
-    match std::fs::write(path_to_repo_metadata, metadata_string) {
-        Ok(_) => (),
-        Err(e) => return status::Custom(
-            Status::InternalServerError,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response(format!("Could not write metadata template to repo: {}", e)),
-            ),
-        )
-    }
-    // Add and commit
-    new_repo.index()
-        .unwrap()
-        .add_all(&["."], git2::IndexAddOption::DEFAULT, None)
-        .unwrap();
-    new_repo.index().unwrap().write().unwrap();
-    let sig = new_repo.signature().unwrap();
-    let tree_id = {
-        let mut index = new_repo.index().unwrap();
-        index.write_tree().unwrap()
-    };
-    let tree = new_repo.find_tree(tree_id).unwrap();
-    new_repo.commit(Some("HEAD"), &sig, &sig, "Initial commit", &tree, &[]).unwrap();
-    status::Custom(
-        Status::Ok,
-        (
-            ContentType::JSON,
-            make_good_json_data_response("ok".to_string()),
         ),
     )
 }
